Replace the "list" componento a single item component (better to use it with any list cmponent I want)
